pyproject.toml
[tool.poetry]
name = "kafka-project"
version = "0.1.0"
description = ""
authors = ["VitorPudo <vitor.silva@shopee.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "2.0.2"
pandas = "^2.2.2"
boto3 = "^1.34.93"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
